requirements.txt
asn1crypto==0.24.0
blurhash==1.1.4
certifi==2024.7.4
cffi==1.14.4
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.2
cryptography==43.0.3
decorator==4.3.0
dparse==0.5.2
http-ece==1.0.5
idna==3.7
Mastodon.py==1.8.1
packaging==17.1
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.3
python-magic==0.4.27
pytz==2018.5
PyYAML==6.0.2
requests==2.32.3
six==1.11.0
toml==0.10.2
urllib3==2.2.2
